+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 * `KrontabTemplateWrapper` has been added
 * New extension `KrontabTemplate#toKronScheduler` (works as `toSchedule`)
 * **Fixed issue related to the fact that `toNearDateTime` of `CronDateTime` incorrectly handled months
-
+* **Android target has been added**
 
 ## 0.4.2